Dispatch R6 Class

Send messages to jamovi tables from anywhere in the code. Warnings can be sent to tables or arrays. For tables in Array, they do not need to exist at the moment the message is broadcast, because the message is stored in the parent object and then passed to the table at run time.

Warnings can be transient (get removed after .init()) when init=TRUE is passed.

Errors must be sent to existing (already defined) objects. Errors are passed directly to the jamovi object. If option final=TRUE, a stop() is issued @export

Active bindings

warnings

Set a warning message to a jamovi table It accepts a list with fields:

topic the nickname of a table in the format:

tablename a name of the table as set in the r.yaml jamovi file

groupname_tablename the group the table belongs to and the table name

arrayname the array of the tables target of the message

message the message to broadcast

init (optional) init=TRUE makes the message transient (disappears after init)

errors

set an error message to a jamovi table It accepts a list with fields:

topic the nickname of a table in the format:

tablename a name of the table as set in the r.yaml jamovi file

groupname_tablename the group the table belongs to and the table name

arrayname the array of the tables target of the message

message the message to broadcast

final (optional) final=TRUE makes module stop computation immediately

Methods

Public methods

Dispatch$new()

Method new()

Initialize the Dispatch object

Usage

Dispatch$new(results)

Arguments

results

A jamovi results object jmvcore::Results