SmartArray R6 Class

A programmer interface to init and polutate jmvcore::Array

Super class

jmvScaffold::SmartTable -> SmartArray

Public fields

childrenObjs

a list of children jmvScaffold::SmartTable

Methods

Public methods

SmartArray$new()

SmartArray$initTable()

SmartArray$runTable()

SmartArray$retrieveNotes()

SmartArray$setNotes()

Inherited methods

jmvScaffold::SmartTable$ci()

jmvScaffold::SmartTable$setColumnTitle()

Method new()

Initialize the SmartArray object, creating the children tables

Usage

SmartArray$new(array, estimator = NULL)

Arguments

array

A jamovi array of class jmvcore::Array

estimator

optional R6. If set, SmartArray inquires the R6 class for init and run functions to initialize and populate the jamovi tables. the R6 class should provide functions using the convention init_[nickname] and run_[nickname] where the nickname` is the unique identifier given to the jamovi table. The unique identifiers can be

arrayname: the name of the jmvcore::Array

groupname_arrayname: the name of the jmvcore::Group containing the jmvcore::Array and the Array name

Method initTable()

initializes all children tables from runSource or estimator::run_[tablename], and takes care of all aesthetic adjustments required

Usage

SmartArray$initTable()

Method runTable()

fills all children tables from runSource or estimator::run_[tablename], and takes care of all aesthetic adjustments required

Usage

SmartArray$runTable()

Method retrieveNotes()

used internally, but must be kept public

Usage

SmartArray$retrieveNotes()

Method setNotes()

set a note to all children tables. Not necessary if a Dispatch object is used instead

Usage

SmartArray$setNotes(note)

Arguments

note

a string: the message to set